requirements.txt
streamlit==1.31.0
requests==2.31.0
pyyaml==6.0.1
boto3==1.34.0
pandas==2.1.3
sseclient-py==1.7.2
python-dotenv>=1.0.0
uuid==1.30
